Dev testing

+++ scripts/install.sh
@@ -4,10 +4,16 @@
 # Machine Learning Development Server Stack
 #
 # Copyright 2020 Puget Systems and D B Kinghorn
-
-# This script will do some sanity checks for installed software
-# and compatability with the server stack setup configurations
-#
+# 
+# This script will;
+# - check for a compatable base OS
+# - install needed extra software packages
+# - install/update the NVIDIA display driver if needed
+# - Insatll and configure Cockpit
+# - Set netplan to use NetworkManager for Cockpit (if not set)
+# - Install and configure JupyterHub with "single-user notbook server"
+#   using JupyterLab
+# - Add some default kernelspecs for JupyterLab 
 
 set -o pipefail
 
@@ -15,6 +21,15 @@ set -o pipefail
 #trap read debug
 
 SCRIPT_HOME=$(pwd)
+
+CONDA_HOME=/opt/conda
+JHUB_HOME=${CONDA_HOME}/envs/jupyterhub
+JHUB_CONFIG=${JHUB_HOME}/etc/jupyterhub/jupyterhub_config.py
+JUPYTER_SYS_DIR=/usr/local/share/jupyter
+KERNELS_DIR=${JUPYTER_SYS_DIR}/kernels
+
+NVIDIA_DRIVER_VERSION='450'
+USEGPU=''
 
 ERRORCOLOR=$(tput setaf 1)    # Red
 SUCCESSCOLOR=$(tput setaf 2)  # Green
@@ -35,24 +50,58 @@ source /etc/os-release
 if [[ $NAME == "Ubuntu" ]] && [[ ${VERSION_ID/./}+0 -ge 1804 ]]; then
     success "[OK] ${PRETTY_NAME}";
 else
-    error "OS must be Ubuntu 18.04 or greater"
+    error "[STOP] OS must be Ubuntu 18.04 or greater"
     exit 1
 fi
 
+# Install extra packages (may already be installed)
+note "Installing extra packages -- curl openssl build-essential dkms emacs-nox"
+apt-get install --yes -qq curl openssl build-essential dkms emacs-nox
+
 note "Checking for NVIDIA GPU"
 
-[[ $(lspci | grep NVIDIA) ]] && success "[OK] Found NVIDIA GPU" || error "[Warning] NVIDIA GPU not detected, using CPU-only install..."
+function add_nv_driver() {
+    # Args: "driver version"
+    #apt-get install -q dkms
+    add-apt-repository --yes -q ppa:graphics-drivers/ppa
+    apt-get update
+    apt-get install --no-install-recommends --yes -q nvidia-driver-$1
 
-if [[ $(which nvidia-smi) ]]; then 
-    driver-version=$(nvidia-smi | grep Driver | cut -d " " -f 3) 
-    note "Driver Version = ${driver-version}"
-    if [[ ${driver-version%.*}+0 -lt 440 ]]; then
-        error "Your NVIDIA Driver is out of date! ... Updating"
-        #add driver update
+    sudo tee -a /etc/modprobe.d/blacklist-nouveau.conf << 'EOF'
+blacklist nouveau
+options nouveau modeset=0
+EOF
+    update-initramfs -u
+}
+
+function get_driver_version() {
+     nvidia-smi | grep Driver | cut -d " " -f 3;
+}
+
+#function add_nv_driver() {
+#    # Args: driver version 
+#    sudo apt-get install --yes -q nvidia-driver-$1-server
+#}
+
+if lspci | grep -q NVIDIA; then
+    success "[OK] Found NVIDIA GPU"
+    USEGPU=0 # 0=true
+    if [[ $(which nvidia-smi) ]]; then
+        driver_version=$(get_driver_version)
+        note "Driver Version = ${driver_version}"
+        if [[ ${driver_version%%.*}+0 -lt ${NVIDIA_DRIVER_VERSION} ]]; then
+            error "Your NVIDIA Driver is out of date! ... Updating"
+            add_nv_driver ${NVIDIA_DRIVER_VERSION} && success "[OK] NVIDIA Driver Installed" \
+                || error "!!Driver install failed!!"
+        fi
+    else
+        error "[Warning] NVIDIA Driver not installed ... Installing now"
+        add_nv_driver ${NVIDIA_DRIVER_VERSION} && success "[OK] NVIDIA Driver Installed" \
+               || error "!!NVIDIA Driver install failed!!"
     fi
 else
-    error "[Warning] NVIDIA Driver not installed" 
-    # install driver
+    error "[Warning] NVIDIA GPU not detected, using CPU-only install..."
+    USEGPU=1 # 1=false
 fi
 
 #
@@ -119,13 +168,18 @@ systemctl restart direvent
 # Cockpit messes up if it is not using NetworkManger so lets change Ubuntu server to use it
 #
 
-note 'Changing netplan to use NetowrkManager on all interfaces ...'
-# backup existing yaml file
-cd /etc/netplan
-cp 01-netcfg.yaml 01-netcfg.yaml.BAK
+note "Checking netplan configuration ..."
+if grep -q -E 'renderer.*NetworkManager' /etc/netplan/*.yaml; then
+    note "[OK] NetworkManager in use"
+else
+    note 'Changing netplan to use NetowrkManager on all interfaces ...'
+    # backup existing yaml file
+    cd /etc/netplan
+    note "backing up existing yaml files"
+    for i in *.yaml; do mv "$i" "$i.bak"; done
 
-# re-write the yaml file
-sudo tee /etc/netplan/01-netcfg.yaml << 'EOF'
+    # re-write the yaml file
+    sudo tee /etc/netplan/01-netcfg.yaml << 'EOF'
 # This file describes the network interfaces available on your system
 # For more information, see netplan(5).
 network:
@@ -133,10 +187,14 @@ network:
   renderer: NetworkManager
 EOF
 
+fi
+
 # setup netplan for NM
+note "Activating NetworkManager netplan"
 netplan generate
 netplan apply
 # make sure NM is running
+note "Enabeling/starting NetworkManager service"
 systemctl enable NetworkManager.service
 systemctl restart NetworkManager.service
 
@@ -147,36 +205,29 @@ success "[OK] Cockpit installed"
 #
 note "Installing JupyterHub ..."
 
-# 
-
-CONDA_HOME=/opt/conda
-JHUB_HOME=${CONDA_HOME}/envs/jupyterhub
-JHUB_CONFIG=${JHUB_HOME}/etc/jupyterhub/jupyterhub_config.py
-JUPYTER_SYS_DIR=/usr/local/share/jupyter
-KERNELS_DIR=${JUPYTER_SYS_DIR}/kernels
-
-# Install extra packages (may already be installed)
-apt-get install --yes -qq curl openssl build-essential emacs-nox
-
 #
 # Install conda (globally)
 #
 
 # Add repo
+note "...Adding conda/miniconda repo to apt source.list.d/conda.list..."
 curl -L https://repo.anaconda.com/pkgs/misc/gpgkeys/anaconda.asc | apt-key add -
 echo "deb [arch=amd64] https://repo.anaconda.com/pkgs/misc/debrepo/conda stable main" | sudo tee  /etc/apt/sources.list.d/conda.list 
 
 # Install
+note "... Installing conda..."
 apt-get update
 apt-get install --yes -qq conda  
 
 # Setup PATH and Environment for conda on login
+note "...Adding conda init script to profile.d"
 ln -s ${CONDA_HOME}/etc/profile.d/conda.sh /etc/profile.d/conda.sh
 
 # source the conda env for root
 . /etc/profile.d/conda.sh
 
 # Update miniconda packages
+note "...Updating conda, python, utilities..."
 conda update --yes conda
 conda update --yes python
 conda update --yes --all
@@ -186,11 +237,13 @@ conda update --yes --all
 #
 
 # Create conda env for JupyterHub and install it
+note "...Creating jupyterhub env, installing jupyterhub, jupyterlab, ipywidgets..."
 conda create --yes --name jupyterhub  -c conda-forge jupyterhub jupyterlab ipywidgets
 
 # Set highest priority channel to conda-forge
 # to keep conda update from downgrading to anaconda channel
 # This all needs to happen in the jupyterhub env!
+note "...Setting jupyterhub (local) environment to use conda-forge for updates..."
 conda activate jupyterhub
 touch $JHUB_HOME/.condarc 
 conda config --env --prepend channels conda-forge
@@ -201,20 +254,24 @@ conda update --yes --all
 #
 # create and setup jupyterhub config file
 #
+note "...Creating JupyterHub config..."
 mkdir -p ${JHUB_HOME}/etc/jupyterhub
 cd ${JHUB_HOME}/etc/jupyterhub
 ${JHUB_HOME}/bin/jupyterhub --generate-config
 
 # set default to jupyterlab
+note "...Setting default spawner to JupyterLab..."
 sed -i "s/#c\.Spawner\.default_url = ''/c\.Spawner\.default_url = '\/lab'/" jupyterhub_config.py
 
 # don't show the install Python kernel spec
+note "...Allow removal of default python kernelspec..."
 sudo tee -a ${JHUB_HOME}/etc/jupyter/jupyter_notebook_config.py << 'EOF'
 # Allow removal of the default python3 kernelspec
 c.KernelSpecManager.ensure_native_kernel = False
 EOF
 
 # add SSL cert and key for using https to access hub
+note "...Creating SSL certificate (self-signed).."
 mkdir -p ${JHUB_HOME}/etc/jupyterhub/ssl-certs
 cd ${JHUB_HOME}/etc/jupyterhub/ssl-certs
 
@@ -232,6 +289,7 @@ sed -i "s/#c\.JupyterHub\.ssl_key =.*/c\.JupyterHub\.ssl_key = '\/opt\/conda\/en
 #
 
 # Create a systemd "Unit" file for starting jupyterhub,
+note "...Setting up systemd service unit file... "
 mkdir -p ${JHUB_HOME}/etc/systemd
 
 # have to use a regular here doc to parse CONDA_HOME
@@ -253,17 +311,18 @@ EOF
 ln -s ${JHUB_HOME}/etc/systemd/jupyterhub.service /etc/systemd/system/jupyterhub.service
 
 # Start jupyterhub, enable it as a service,
+note "...Starting JupyterHub service..."
 systemctl start jupyterhub.service 
 systemctl enable jupyterhub.service
 
 #
 # Add some extra kernels for JupyterLab
 #
-
+note "...Adding extra default kernelspecs for JupyterLab..."
 # make sure we are in the script dir
 cd ${SCRIPT_HOME}
 
-add_kernel() {
+function add_kernel() {
     # Args: "env-name" "package-name(s)" "display-name" "icon"
     ${CONDA_HOME}/bin/conda create --yes --name $1 $2
     ${CONDA_HOME}/bin/conda install --yes --name $1 ipykernel
@@ -273,11 +332,12 @@ add_kernel() {
     fi 
 }
 
-# Anaconda3
-#add_kernel "anaconda3" "anaconda" "Anaconda3 All" "anacondalogo.png"  
+# Add some kernels by default
+add_kernel "py3" "python=3" "Python 3"
+#add_kernel "anaconda3" "anaconda -c anaconda" "Anaconda Python3" "anacondalogo.png"  
 #add_kernel "tensorflow2-gpu" "tensorflow-gpu" "TensorFlow2 GPU" "tensorflow.png" 
 #add_kernel "tensorflow2-cpu" "tensorflow" "TensorFlow2 CPU" "tensorflow.png" 
-add_kernel "pytorch-gpu" "pytorch torchvision -c pytorch" "PyTorch GPU" "pytorch-logo-light.png" 
+#add_kernel "pytorch-gpu" "pytorch torchvision -c pytorch" "PyTorch GPU" "pytorch-logo-light.png" 
 
 
 #${CONDA_HOME}/bin/conda create --yes -q --name anaconda3 anaconda ipykernel
@@ -287,12 +347,13 @@ add_kernel "pytorch-gpu" "pytorch torchvision -c pytorch" "PyTorch GPU" "pytorch
 #
 # remove the jupyter kernelspec for the system miniconda python3 
 #
+note "...Removing default miniconda python kernelspec ..."
 echo "y" | ${JHUB_HOME}/bin/jupyter kernelspec remove python3 
 
 #
 # Add PSlabs branding to jupyterhub login page
 #
-
+note "...Adding Puget Systems Labs branding to JupyterHub login page..."
 # make sure we are in the script dir
 cd ${SCRIPT_HOME}
 
@@ -314,8 +375,15 @@ ln -s pslabs-login.html login.html
 # will need a restart of jhub before it works right 
 
 # hack for Ubuntu 18.04 ...sudo writes root owned clocal confi file to script runners home dir!
+note "...Clean up Ubuntu 18.04 'broken' sudo rubble from installers home dir... "
 if grep -q 'bionic' /etc/os-release ; then
     $(chown -R  ${SUDO_USER}:${SUDO_USER} ${HOME}/.*)
 fi
+
+success "*****************************************"
+success "INSTALL COMPLETE -- Please Restart System"
+success "Admin interface is on port 9090"
+success "JupyterHub login is on port 8000"
+success "*****************************************"
 
 exit 0
